Remove subdirectory in registry path
We no longer need dbtype-specific registry files. We can use generic registries,
and if we need to specify an additional server for related databases (e.g. a core db if we are testing an otherfeatures db)
that can be handled with the new 'target_url' parameter (which maps to 'server_uri' in the underlying datacheck pipeline).