Hide some job table columns by default.
The datacheck table has quite a lot of columns,
generally necessitating horizontal scrolling, which
is a pain. Add in the datacheck name field, but
make it hidden by default, along with the group and
type - columns can be added by user, and this info
is visible by clicking on the '+' button.
Also deleted a couple of duplicate tags.